+++ lib/toy/instrumentation/active_support_notifications.rb
@@ -1,0 +1,5 @@
+require 'securerandom'
+require 'active_support/notifications'
+
+# Set the instrumenter to active support notifications.
+Toy.instrumenter = ActiveSupport::Notifications