+++ perf/writes.rb
@@ -1,14 +1,11 @@
 # require 'perftools'
 require 'pp'
-require 'logger'
 require 'benchmark'
 require 'rubygems'
 
 $:.unshift File.expand_path(File.dirname(__FILE__) + '/../lib')
 require 'toystore'
 require 'adapter/memory'
-
-Toy.logger = ::Logger.new(STDOUT).tap { |log| log.level = ::Logger::INFO }
 
 class User
   include Toy::Store