+++ spec/helper.rb
@@ -1,12 +1,9 @@
 $:.unshift(File.expand_path('../../lib', __FILE__))
 
 require 'pathname'
-require 'logger'
 
 root_path = Pathname(__FILE__).dirname.join('..').expand_path
 lib_path  = root_path.join('lib')
-log_path  = root_path.join('log')
-log_path.mkpath
 
 require 'rubygems'
 require 'bundler'
@@ -19,15 +16,13 @@ require 'support/objects'
 require 'support/identity_map_matcher'
 require 'support/name_and_number_key_factory'
 require 'support/shared_active_model_lint'
-
-Logger.new(log_path.join('test.log')).tap do |log|
-  Toy.logger = log
-end
+require 'support/instrumenter_helpers'
 
 RSpec.configure do |c|
   c.include(Support::Constants)
   c.include(Support::Objects)
   c.include(IdentityMapMatcher)
+  c.include(InstrumenterHelpers)
 
   c.fail_fast = true
   c.filter_run :focused => true
@@ -38,5 +33,6 @@ RSpec.configure do |c|
   c.before(:each) do
     Toy::IdentityMap.enabled = false
     Toy.key_factory = nil
+    clear_instrumenter
   end
 end